+++ .stylelintrc
@@ -5,6 +5,7 @@
     "at-rule-no-unknown": null,
     "no-descending-specificity": null,
     "no-duplicate-selectors": null,
+    "length-zero-no-unit": null,
     "selector-pseudo-class-no-unknown": [
       true, { ignorePseudoClasses: ["global"] }
     ]
