+++ styleguide.webpack.js
@@ -9,7 +9,8 @@ module.exports = {
     plugins: [ new DirectoryNamedWebpackPlugin(true) ],
     alias: {
       components: path.resolve(__dirname, 'src/components/'),
-      styles: path.resolve(__dirname, 'src/styles/')
+      styles: path.resolve(__dirname, 'src/styles/'),
+      utils: path.resolve(__dirname, 'src/utils/')
     }
   },
   node: { fs: 'empty', net: 'empty' },
