+++ webpack.config.js
@@ -52,7 +52,8 @@ const config = {
     plugins: [ new DirectoryNamedWebpackPlugin(true) ],
     alias: {
       components: path.resolve(__dirname, 'src/components/'),
-      styles: path.resolve(__dirname, 'src/styles/')
+      styles: path.resolve(__dirname, 'src/styles/'),
+      utils: path.resolve(__dirname, 'src/utils/')
     }
   },
   optimization: {
